Removed blur from edge detection as it had bad accuracy
For low resolution images, blur was losing too much accuracy for the Sobel Operator to work. So we removed this.

+++ src/gui/FilterParameters.h
@@ -126,7 +126,7 @@ public:
                 show = false;
                 init = false;
                 textureNeedsUpdate = true;
-                gPresetManager.recordStep(FilterStep{FilterType::Blur, { kernal, sigma }, ""});
+                gPresetManager.recordStep(FilterStep{FilterType::Blur, { static_cast<double>(kernal), static_cast<double>(sigma) }, ""});
                 clearStoredOriginalImage();
             }
